chore: refactor and cleanup tests
